requirements.txt
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.9.1
certifi==2025.8.3
charset-normalizer==3.4.3
Cython==3.1.3
dartsclone==0.9.0
distro==1.9.0
Django==5.2.5
django-rest-framework-mongoengine==3.4.1
djangorestframework==3.16.1
dnspython==2.7.0
greenlet==3.2.4
groq==0.31.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-core==0.3.74
langchain-groq==0.3.7
langchain-text-splitters==0.3.9
langsmith==0.4.14
mongoengine==0.29.1
openai==1.99.9
orjson==3.11.2
packaging==25.0
pydantic==2.11.7
pydantic_core==2.33.2
pygtrie==2.5.0
pymongo==4.14.0
python-dotenv==1.1.1
PyYAML==6.0.2
requests==2.32.4
requests-toolbelt==1.0.0
sniffio==1.3.1
sortedcontainers==2.1.0
SQLAlchemy==2.0.43
sqlparse==0.5.3
SudachiDict-core==20250515
SudachiPy==0.5.4
tenacity==9.1.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
urllib3==2.5.0
zstandard==0.23.0
